Avoid a timeout in the test.
Before this change, the way the queue was artificially frozen caused the
replay mechanism to not get back what it send, which caused it to time
out before eventually succeeding.

This change modifies the way the queue is frozen so as not to interfere
with replays.

This change also avoid attempting to go to a cursor position that
doesn't exist in the term buffer (yet), which would also cause a
timeout.