Optimize the replay of simple insertions.
When text is inserted right after some change being replayed, before
this change, the first replay would end, the cursor would be moved to
the cursor and then a second replay would run which might also need to
move the cursor.

With this change, this common case is optimized for the common case of
inserted text: that text is just appended to the last modification. If
the modification has already been sent to the terminal, but hasn't been
echoed back yet, the text is sent to the terminal and the wait condition
modified to wait for the full text.