Introduce a hook to react to interactive text insertions.
Completion UI that appears automatically is usually implemented by
tracking use of self-insert-command and looking at the inserted text in
a post-command hook.

This doesn't work with MisTTY, as typed characters are sent directly to
the terminal to be echoed back, or not, whenever and however the program
attached to the terminal wants.

This change provides a hook that reports that characters were typed
after they've been echoed back. This should make it possible to
implement bridges between MisTTY and completion UIs.

This is related to issue #5; corfu-auto is an example of such a UI.