Document mistty-interactive-insert-hook.
This change adds a section that explains how to configure autocomplete
UI to show up inside of MisTTY:

https://mistty.readthedocs.io/en/latest/usage.html#emacs-autocomplete

This is all a bit hackish and very experimental, which is why the actual
bridges aren't included into the distribution.

Fixes #5 with this latest change and the configuration in the doc,
corfu-auto seems to work.