mistty--cursor-to-point-interaction bails out unless it is last.
cursor-to-point is meant for an interactive scenario: it moves to the
point as it is now, not the point as it was when the interaction was
queued. It's never useful to run this interaction between two other
interactions.

With this change, mistty--cursor-to-point-interaction only works when it
is last on the queue when it is started.