Check overlays in pre-command hook as well.
In the case of an autocompletion UI started from
mistty-interactive-insert-hook on a timer, as is usually the case, we'll
only notice the new overlay after executing the next command.

With this change, the new overlay will be noticed in the pre-command
hook, and most mistty-specific commands will fall back to their Emacs
equivalent.

This is just a quick and dirty workaround, waiting for a proper fix.