Let Emacs track point and window point normally.
This change removes mistty--recenter, which was called at every refresh.

mistty--recenter was meant to keep the window visible, as there were
cases when the prompt would just disappear when windows were resized.
This was very early on in the development. These case seem to have
disappeared

This change resets back to the default Emacs behavior, with just
scroll-conservatively and window-point-insertion-type.

It is possible that cases where the point isn't visible come back and
require some adaptation - hopefully nothing as aggressive as
mistty--recenter. At this point, it doesn't look like it.

Fixes #6

Also relates to issue #7 since mistty--recenter was an unnecessary
source of slowdown.