Use replace-buffer-content only when on a prompt.
replace-buffer-content is expensive and only useful when editing a
command line. See issue #7. It's unnecessary slow when just displaying
command output.

This change switches to using copy-buffer-substring outside of a prompt.
This speeds up long command output noticeably.

This required some adaptation to the tests and to the sync mark history,
which were relying on markers and narrow-region being conserved within
the synced region. This isn't the case anymore when
copy-buffer-substring is used.