Avoid autosuggestions confusing completion-at-point.
Starting with this change, MisTTY wraps functions in
completion-at-point-functions to prevent them from seeing
autosuggestions, or, to be more exact, to prevent them from seeing
what's after the point. They complete the text before the point instead
of the text around the point.

This can be turned off by setting the option mistty-wrap-capf-functions
to nil.