Replay: fix and apply upper-limit.
Before this change, upper-limit was never set. What made the test pass
was the fact that insertion ended up being empty, as the point was too
far from orig-beg. When there was only one change, it didn't matter, but
when there was many (a global search-and-replace), this took forever, as
each change required 2 timeouts: one to fail to go to beg and one to
fail to go to old-end.

With this change, upper-limit is set, the modification is skipped, and
then all modifications after that that start to the right of upper-limit
are skipped as well.

A new test variable, mistty-expected-issues, helps make sure that
there's only one timeout in test.