replay: insert everything or not at all.
Before this change, when inserting, if replay couldn't go back to the
original insertion point, it tried to insert the substring that
corresponded to the difference between the original insertion point and
the point it could reach.

This behavior was more confusing than helpful.

With this change replay inserts either everything, if the position
reached was after the beginning (this commit), or not at all if the
position reached was before the beginning (previous commit).