Simplify and shorten replay, test replacements at prompt.
Before this change, replay would:
- go to the beginning of the change,
- check the position, then
- go to the start position for deletion (old-end),
- check the position, then
- do delete and insert in one go.

Before this change, replay would:
- goes to the start position for delete and insert,
- check the position, then
- do delete and insert in one go.

The previous scheme was meant to know where to shorten insertion and
deletions, but MisTTY doesn't do that anymore. It applies insertion and
deletion fully or not at all.

This change also updates the code that deals with lower limit so that
once a lower limit has been detected, replay stops trying to execute
changes.

New tests cover the cases where replacement or deletion is dropped at
prompt. The tests also make sure that there's only one timeout, to make
sure that lower-limit is correctly applied.