Always keep point after a replay.
In normal operation, if the terminal content is refreshed and the point
was at cursor before the refresh, mistty--refresh moves the point to
follow the cursor. This is convenient when the application tied to the
terminal moves the cursor/point.

When replaying changes made in Emacs, however, following the cursor is
never the right thing, because Emacs is in control. The cursor should
follow the point.

Before this change, there were cases where the point would be at cursor
before a replay, and then, after the replay, refresh would think that
the point should follow the cursor. This change turns off this very
specific case.