Fixes for mistty--self-insert-line.
This change uses reverse instead of nreverse when reading from
mistty--self-insert-line, as nreverse would change the content of the
list, causing strange errors.

This change also moves the update of mistty--self-insert-line inside of
the interaction, so that the point makes sense even if insertion is
delayed.

This change also updates the description of the content of
mistty--self-insert-line.