Strictly restrict overlays seen as long-running commands.
**This is a backward-incompatible change!**

Before this change, unknown overlays would be seen as a sign that a
long-running command is in effect and would turn MisTTY off.

This looked convenient, because it made MisTTY work with other packages
without having to know about them. However, in practice, the resulting
behavior confusing, as it looks like MisTTY just turns itself off for no
apparent reasons.

With this change, only overlays with a known property trigger the
functionality. On one hand, this requires putting keeping track of a
list of known long-running commands that use overlays for these to work
properly in MisTTY. This is annoying, but less dangerous and confusing.

This means that unknown feature that was correctly identified by MisTTY
as a long-running command until it is added to
mistty-foreign-overlay-properties.

With this change MisTTY only identifies the following as long-running
command:
 cua-rectangle
 yas
With support for tempel to be added in a follow-up change.