Doc: better list supported integrations.
This change reworks the documentation to have the Usage section mention
that corfu, company, yasnippet and tempel should work with MisTTY and
move into the discussion of the different issues and solutions that
might be needed for integrating with other packages into the Extension
section.

This change also mentions tide and powerline.