Mark a development version with .9 after a release.
This change introduces the idea of marking a development version by
adding a .9 to the end of the release that preceded it, so the
development version following release 1.0 will be 1.0.9 and the next
release might be 1.1 or 2.0.

While released versions are fixed and tagged in the repository,
development version are changeable; you really need the git or melpa
version to know what you're working with.