Refactor the way interactions are defined.
With this change, mistty--interact-return and mistty--interact-done
actually return from the current function. This makes it easier to write
callbacks for interactions, as they don't need to be crafted so
carefully to return the values to be sent to the terminal. To be safe,
callbacks are not allowed to actually return. They must call one of the
nonlocal return functions.

This change also introduces a macro that make it simpler to build
straightforward interactions with just one callback. The body of the
macro is the body of the callback and the return value of the macro is
the interaction, ready to be enqueued.

This also moves initial-buf and buf initialization into
mistty--make-interact instead of mistty--interact-init, which the new
macro replaces.