Allow some simple modifications even edits are forbidden.
Before this change, if the terminal region matched
mistty-forbid-edit-regexp, any modification in that region was forbidden
and resulted in an error.

This change instead assumes that when edits are forbidden, what's
actually forbidden are cursor movements; you can still type and delete
characters - this is at least what's happening when shells are in
history search mode.

This allows some simple modifications to be replayed.

With this change, when mistty-forbid-edit-regexp matches, MisTTY does as
much as it can without moving the cursor. If the change cannot be
replayed, it is eventually overwritten just like any other unreplayable
change.

This allows simple commands like backward-kill-word or yank to work
while the shell is in history search mode.