Temporarily disable yas/fish tests on CI.
This change disables the following tests:
  mistty-compat-yas-expand-multiline-fish
  mistty-compat-yas-expand-multiline-fish-insert
that are currently failing, presumably due to a version change within
yas. This allows CI passing while investigating the version
incompatibility.

This change adds the option --ci to eldev test, so tests can be skipped
only when run for continuous integration and uses this new option in the
github action workflow.