CI: Tweak CI.yml so act can run it.
Before this change, CI.yml relied on Homebrew being available to install
zsh and fish and just assumed that vi was available. While this is true
on github, this isn't true in the default image used by
https://github.com/nektos/act to run workflows.

This change switches CI.yml to rely only on Nix, installed by
emacs-setup, to install everything needed to run the tests and generate
the documentation.