Skip empty modifications.
Before this change, if the change detected was "delete until the end of
the line" and there was nothing to delete, the cursor was unnecessarily
moved to the end to get ready for the empty deletion, which could time
out when the end was not accessible.

This issue caused the following tests to fail with the latest yasnippet
release:
 mistty-compat-yas-expand-multiline-fish
 mistty-compat-yas-expand-multiline-fish-insert

With this change, such no-op changes are skipped, which avoids an
unnecessary time-out during normal use and an error during testing.

This reverts change b729113