Limit right prompt detection logic to the prompt.
Before this change, output containing tabs or otherwise skip characters
would trigger the right prompt detection logic, which would then attempt
to remove from the yanked text what it sees as part of the prompt.

This change restricts right prompt detection to the prompt, the section
with bracketed-paste turned on, so that command output is left alone.

Fixes #21