fix: mistty-kill-buffer(-and-window) do nothing on nonzero exit code
Before this change, mistty-kill-buffer(-and-window) would only test that
the process exited, but not its status code. This had surprising
results, such as when explicit-shell-file-name was not set properly as
the buffer would be killed before the user could see the error message.

With this change, mistty-kill-buffer(-and-window) only kill the buffer
if the process exited successfully.