feat: Execute remote shells with TRAMP.
Before this change, when default-directory of the current buffer was a
remote file, mistty-create would create a new shell in the user's home
directory, like term does.

With this change, in such situation, mistty-create creates a new shell
on the corresponding remote host, if TRAMP supports it.

This change also extends mistty-create to accept a command with
arguments, specified as a list, instead of just a single command.