fix: Avoid TRAMP freezing when executing a shell on MacOS.
Before this change, starting a process on a mac through TRAMP would just
freeze. This change works around that issue.

On MacOS, the length of the termcap entry, heavily escaped by TRAMP,
plus the other env variables, when passed to the shell using the env
command, is enough to hit the 1024 byte limit of the tty cache used in
canonical mode. On Linux, that limit is 4095, according to termios(3),
so there are no problems.

Adding a newline to the termcap entry is enough to avoid hitting that
limit while keeping the termcap entry valid. An alternative would be to
have TRAMP disable canonical mode with stty -icanon before sending out
the command.