feat: Introduce custom option mistty-shell-command
Before this change, to configure the shell MisTTY runs, you'd set
explicit-shell-file-name. This allows setting a single executable only.
shell.el allows setting arguments by setting explicit-<shellname>-args,
but MisTTY doesn't use that, as it typically contains arguments that
disable command-line editing, such --noeditor for bash.

There was then no way of having MisTTY sets specific arguments to the
shell it runs, short of calling mistty-create directly.

With this change, it is possible to set the new customization
option mistty-shell-command to define the command for mistty and
the default command for mistty-create.

mistty-shell-command can also be configured as a connection-local
variable to set different commands for different machines.