feat: (mistty) creates new buffer with the original default-directory.
Before this change, when mistty created a new buffer, it could be based
on the default-directory of the last mistty buffer that was visited.

With this change, mistty remembers the buffer that was current the first
time a mistty command in a chain was called and that buffer's
default-directory is used if the mistty command eventually decides to
create a new buffer.

This is especially useful when the buffer TRAMP host doesn't match the
TRAMP host of the existing mistty buffer.