fix: When the shell is remote interpret \032 <path> \n as remote.
Before this change, when Bash was run remotely, it would send out its
current directory to Emacs as \032 <path> \n. term-command-hook would
then try to call cd on that a local path and fail, because that path is
only valid on the remote host.

This change interprets paths sent as \032 <path> \n as being local to
the remote host set by default-directory. This allows the use-case of
Bash running remotely to just work out of the box without extra
configuration.