feat(mistty-osc7): Build TRAMP paths relative to default-directory.
Before this change, when running the shell as sudo with TRAMP sudo
method, mistty-osc7 would build local paths as default-directory,
erasing the TRAMP prefix.

With this change, in such a case, mistty-osc7 builds paths using the
TRAMP sudo method - that is, using the same TRAMP prefix as
default-directory.

This is generalized in the implementation as "if the hostname in the
file: URL correspond to the hostname of the current TRAMP prefix, keep
that prefix unmodified."