fix: Remove fake newlines outside of the terminal zone.
Before this change C-e could behave strangely in the scrollback zone if
the window was resized after the command was executed. This was due to
fake newlines that term.el adds remaining there, though tagged with the
invisible property.

This change avoids such issue by removing fake newlines when a region of
text transitions from the terminal to the scrollback zone. Fake newlines
are kept in the terminal zone, to keep the terminal zone consistent with
the content of the term-mode buffer.

Unfortunately, deleting fake newlines makes testing much more difficult,
as positions tend to become invalid. To keep tests sane, the new
behavior is suppressed in most tests for now.

Fixes issue #24