feat: Make buffer name configurable, include user and host.
This change introduce the customization option mistty-buffer-name, which
allows configuring what a new MisTTY buffer is going to be called.

The default value of mistty-buffer-name includes function that'll
include the user and host of a remote shell started with TRAMP - or just
the user when using something like the sudo method. That is, with this
change, mistty buffers might have names like *mistty@example.com*
or *mistty-root*.