feat: Build TRAMP paths from alist even if generally disabled.
This change subtly modifies the interaction between
mistty-host-to-tramp-path-alist and mistty-allow-tramp-paths.

Before this change, if mistty-allow-tramp-paths was nil, the values in
mistty-host-to-tramp-path-alist didn't matter; they were just ignored.
With this change, if a path is configured in
mistty-host-to-tramp-path-alist, a TRAMP path is generated, even if
mistty-allow-tramp-paths is nil.

The new behavior allows limiting TRAMP path generation to a subset of
known hosts.