fix: Allow remapping up/down/... in fullscreen mode.
Before this change, remapping keys in mistty-term-key-map would not
apply to keys sent in fullscreen mode, as these are hardcoded in
term.el.

After this change change, all keys defined in raw mode term.el to send
to the terminal to mistty-send-key, which applies mistty-term-key-map.

With the default mistty-term-key-map, this change should not modify what
is sent out to the terminal for any of the affected keys, as
mistty-term-key-map is consistent with term.el.

This change also mirrors some mapping of mistty-mode-map to
mistty-fullscreen-map, for consistency.