fix: Restore original cursor after hiding it.
Before this change, MisTTY would always set cursor-type to t after
hiding it, possibly changing the cursor type set globally or configured
in the current buffer.

With this change, the local value of cursor-type is saved before it is
set to nil and then restored. If there is no local value, cursor-type is
made global again when showing the cursor again.

This fixes #25