fix: Set buffer mode before calling pop-to-buffer.
Before this change, pop-to-buffer was called on a new MisTTY buffer
before its major mode was assigned. This made it impossible to configure
how new MisTTY buffer were displayed in display-buffer-alist by checking
the major mode.

With this change, setting the mode and executing the command is split
into two steps. This way the mode is set to mistty-mode before
pop-to-buffer is called, while pop-to-buffer remains before
mistty--exec, so the terminal window is known when the command is
started.