feat: mistty-set-EMACS controls whether to set the EMACS env var.
Before this change, mistty would set the EMACS env variable if the
locally running version of bash is <= 4.3 (inherited from term.el). This
made no sense when not using Bash as a shell or when connecting to a
remote machine.

With this change, the EMACS env variable is only set if mistty-set-EMACS
is set. That is, it must be turned on by the user. mistty-set-EMACS can
be set as connection-local variable, so can be turned on or off
depending on the host TRAMP connects to.

If running on a host with an older version of bash installed, such as on
macos, this change means that the EMACS env variable is not set anymore
and mistty-set-EMACS must then be set to t to fix the issue.