feat: Allow the default colors to change as the theme changes.
Before this change, terminal colors were always hardcoded in the
face-font-lock text properties. This meant that, when changing theme,
the color would change everywhere except the terminal.

With this change, the default background and foreground colors are left
unspecified in the face-font-lock text properties. This solves the
common case, where the default colors are used.

Non-default colors remain hardcoded, however. These still won't change
as the theme changes, even if the underlying face changed (if, for
example, the text was specified to be "ansi blue" and the new face
specifies another RGB value for blue, the old RGB value is still used.)