fix: Fix whole-line deletion with right prompt.
Before this change, when asked to delete the whole prompt line, usually
to the end of the buffer (marked -1 in the modifications), MisTTY would
take a right prompt into account as text to delete and then fail because
it cannot move the cursor there.

With this change, when asked to delete a whole line, mistty ignores any
final text marked 'mistty-skip as well as the final \n, because it knows
these cannot be deleted.