feat: Scroll up to show completion below point, if necessary.
Before this change, if the point was at the bottom of the window,
completion could appear below the end of the window. Emacs wouldn't see
anything wrong with that, because the point was still visible.

This change detects the case where text appears below the point that is
at the bottom of the window and scrolls the window up to show it.