fix: Keep the mark active when moving the cursor on the prompt
Before this change, mistty-goto-cursor deactivated the mark before
moving the cursor while on the prompt. The idea was that having the mark
expand as the terminal changes would have surprising result.

This isn't convenient when we actually want the terminal to move the
point, such as when sending C-a or C-e to the shell. It's very natural
to do C-SPC C-e to get a copy of the end of the command, but that won't
work if the mark is deactivated. In such cases, deactivating the mark
comes at a bad surprise

With this change, the mark is only deactivated when sending a
command (RET).