test: Clear EMACS env variable before tests that check it.
Before this change, mistty-test-set-EMACS and
mistty-tramp-test-set-EMACS, trying to test that EMACS is not set by
term.el, would assume that the env variable was unset to begin with.
This might not be the case, for example, when running Eldev with a
specific version of emacs, so these tests failed in such case.

This change clears the env variable before testing it.