test: Define and use a dummy TRAMP method for the tests.
Before this change, MisTTY TRAMP tests used the method "sg" as protocol,
which allows running tests with requiring a remote connection or a
password. This didn't work under MacOS, since the command "sg" isn't
available there. It appeared to work under 29.1, but that was because
sg failing was just ignored.

This change switches to a truly dummy TRAMP method that just calls
/bin/sh for the tests, which should work everywhere.