refactor: Turn mistty--cursor-after-last-refresh into a marker.
Before this change, mistty--cursor-after-last-refresh was a position in
the buffer. Such a position could be invalidated if anything before it
was deleted.

Additionally, mistty--cursor-after-last-refresh was tested after the
refresh, but it might have been too late, then, as the refresh could
have invalidated the position. This change checks before the refresh and
just remembers whether the point was at the last known cursor position.

This should make cursor tracking more reliable in corner cases, such as
when a reset happens.