fix: Fix reset behavior, avoiding reset loops.
Before this change, MisTTY would sometimes enter into reset loops,
ending with the entire screen being repeated in the work buffer and
cursor tracking being lost, notably when using apt.

This was due to the sync history not choosing the sync point carefully
enough - or indeed the sync history not realizing that there was no good
sync point on the term buffer past term-home-marker.

Additionally, sync history was clunky and of questionable use with the
latest changes, which modify the portion of the work buffer before the
sync mark after the sync mark is moved.

This change fixes the problem by removing the sync history completely
for now.