feat: Re-align the work and term buffers after a reset.
With this change, when a reset is required because of a change in the
terminal above the sync mark, MisTTY attempts to find a position in the
work buffer that corresponds to term-home-marker by looking for the text
*before* term-home-marker in the work log. The idea is that that text
cannot be changed by the terminal anymore, so should reliably be found
in the work buffer, no matter how much or how little the content term
buffer has changed.

The effect of this change should be very similar to the previous
history-based solution, but hopefully more robust.