fix: Fix issue with reverse i-search on bash.
Before this change, a reverse i-search on bash that was immediately
accepted with RETURN would cause a very strange display.

The problem seems to come from term.el, which adds a fake nl just before
deleting chars as instructed by bash with "^[[24P". This causes a fake
nl to be created that is *not* at the end of the visible line (since 24
chars have been deleted since then), which then says when it should
becomes not-fake as a result of a "\n" being sent later.

This is wrong, but the problem isn't visible in a term buffer, only in a
mistty buffer, since mistty hides fake nl.

This might need to be fixed in term.el. This requires more
investigation; This change is just a workaround for mistty.

Fixes #27