fix: Be stricter when setting mistty-skip.
Before this change, mistty-skip and mistty-right-prompt were set in
mistty-term.el by a combination of move-to-column and after-change hook.
This produced very surprising results when running ipython, which tends
to use move-to-column much more than other programs.

With this change, move-to-column set mistty-maybe-skip, which is then
turned into mistty-skip after looking at the whole line to mark:
 - intend (whitespaces at the beginning of the line)
 - trailing spaces (whitespaces at the end of the line)
 - right prompt (non-ws at the right margin of the window)

This way, mistty-skip shouldn't trigger in the middle of a line as it
used to when using ipython, in Issue #28