fix: Fix right-prompt and yank properties on scrollback.
Before this change, right-prompt and yank properties might not be
available on scrollback, depending on the timing of how the shell
refreshes the input line after the user pressed return.

This change attempts to fix that by delaying
mistty-bracketed-paste (which is used for on-prompt) to nil until after
a refresh *but* mistty-bracketed-paste becomes true right away. This is
hackish and confusing and will need more work.