refactor: Move prepare-term-for-refresh into emulate-terminal.
Before this change, mistty-skip text properties were set from
prepare-term-for-refresh, which was called from mistty just before a
complete refresh.

This caused timing issues, because a refresh tends to be quick at the
end of a prompt, which is fine since the prompt is closed.

This change moves the processing to the end of emulate-terminal. This
will cause it to process more incomplete lines. To reduce the impact,
the way changes are tracked is improved to avoid unnecessary processing.