feat: Kill buffer after process end by default (customizable)
Before this change, to kill buffer after process exists, it was
necessary to add a hook to mistty-after-process-end-hook.

With this change, this is done by default and can be controlled by the
custom option mistty-at-end. The old hooks still work; they take
precedence.

To make this more safe enough, this change make sure that
mistty-kill-buffer and mistty-kill-buffer-and-window do nothing in
corner cases where starting the shell failed immediately, otherwise no
error message would be visible. To do that, this change requires the
user to have sent either a return or a C-d at least once to the buffer.

This change also cleans up the test. Now the fullscreen and
after-process hooks are set to nil by default in mistty-testing.el, to
avoid custom configuration interfering with the tests.