feat: (mistty) reuses dead MisTTY buffer with the same name.
With this change, if MisTTY wants to call a new buffer *mistty*<2>, and
there's already a *mistty*<2> buffer whose process is dead, it'll get
rid of that buffer and reuse the name instead of calling the new
buffer *mistty*<3>.

The new behavior can be turned off by setting the customize option
mistty-force-reuse-buffer-name to nil.

Fix issue #30.