fix: Avoid moving to or even computing unnecessary cursor move.
Before this change, with mistty-skip-empty-spaces set and the pointer at
the corder of a mistty-skip section, MisTTY would decide to move the
point to the current position and report it into mistty-log every time
the pre-display hook was called.

With this change, MisTTY avoids computing the mistty-skip regions unless
the cursor was moved and avoids moving the cursor to its current
position.

Fixes #35