fix: Delete an unknown number of trailing ws with C-k.
Before this change, some spaces could be left at the end of the prompt
that's meant to be completely deleted, because we can't tell whether
these spaces are really there in Zsh (Issue #34).

To avoid this, this change issues a C-k to delete an unknown number of
trailing spaces at the end of the prompt.

Follow-up to a4b6164