fix: Move up/down lines with C-e/C-a instead of counting chars.
Before this change, MisTTY would move through lines with just left and
right. This relied on knowing exactly how many cursor moves to send to
get to the goal.

This approach doesn't work on Zsh, which doesn't always clear spaces
that have been deleted, so they appear as trailing spaces to MisTTY when
in reality, they're not reachable (Issue #34)

This change works around the issue by not relying on knowing exactly the
number of characters to skip to move through lines. This new approach
uses C-a+left and C-e+right to move vertically, look where the cursor
ended up and then issue left/right commands to move horizontally.

Using this C-a/C-e works hard to avoid using up/down, because it can
trigger a history search. This works well for zsh and fish. However,
this relies on C-a/C-e not having been remapped to something else by the
user.

Using C-a+left or C-e+right to move lines doesn't work on bash, which
consider a multiline prompt one line still, so for bash this approach
still relies on horizontal movements and correctly interpreting spaces.