fix: Remove timeout when inserting \n in an existing string.
Before this change, inserting a \n with shirt-RET in the middle of a
line in bash and zsh would cause a hard timeout, because spaces would
then be written to delete the previous text, which should be ignored.

This change ignores such spaces before an inserted newline, avoiding
that timeout.

Note that when interacting manually, such timeout likely wouldn't have
be visible.