fix: Deal with spaces left after splitting a line in bash.
When splitting a line in the middle using shift-RET in bash and zsh, the
shell overwrites characters after the split with spaces. There spaces
should be skipped when computing positions.

Before this change, these spaces could confuse bash multiline movements
for bash, since bash still relies on horizontal-only moves. This didn't
affect moves and replay zsh.

With this change, these spaces are detected and skipped properly.

Note that the effect of this issue was very transient, as bash would
rather quickly overwrite the dead spaces and properly end the line. It
could still confuse replay, however, so we're better off with this fix.