fix: Skip dead spaces with mistty-skip-empty-spaces.
In some situations, Zsh overwrites characters that are deleted with
spaces. This results in spaces that appear to exist, but aren't
reachable on the terminal.

Before this change, it would be possible to reach these spaces even with
mistty-skip-empty-spaces.

With this change, these spaces aren't reachable either, as long as
they're detected (only the case where a \n is inserted in the middle of
a line is currently detected.)