docs: Mention possibly leftover trailing spaces in multi-line.
This change mentions the fact that Emacs commands might leave trailing
spaces when editing multi-line commands with Bash or Zsh.

As discussed in issue #34, MisTTY cannot safely delete these spaces
because it cannot be sure whether they're real or not. A workaround
exists, but it's only safe to use for the last line.