fix: Execute insert even if we can't delete as much as we wanted.
Before this change, if MisTTY failed to delete as far as it wanted, that
is, if the prompt ended earlier than MisTTY thought, the whole
modification would be dropped.

With this change, it'll delete as much as it can and still insert as
long as it managed to put the pointer after beg.