fix: Never move point at cursor.
This change disables mistty-cursor-skip in the case where the point
position corresponds to the last known cursor position. This is always
correct: the cursor position is always a valid position no matter what.

Before this change, it might be possible for an intermediate position to
be rejected by mistty-cursor-skip, resulting in the cursor moving,
seemingly as a result of user interaction, but really just because of
mistty-cursor-skip.

See issue #37