fix: Fix issue with previous-output with zsh completion.
Before this change, previous-output could jump to the end of a previous
line where zsh completion was triggered.

The issue was cause by holes when setting the text property
mistty-input-id, which previous-output uses to detect prompt start and
end.

mistty-register-text-properties sets properties using an
after-change-hook to set properties on text that has been inserted
between the time a terminal feature has been turned on and the time it
is turned off.

This doesn't work in the case where the section that should be covered
contains existing spaces or newline that were added by previous move
operations. In such a case, the terminal doesn't actually need to insert
these, even though they're part of the section.

This issue impacts the property mistty-input-id when there's already a
final newline in the buffer, term.el doesn't need to insert another
newline, which means that the first newline in a mistty-input-id might
not have this properties, causing strange behavior in
previous/next-output.

To work around that, this change has mistty-register-text-properties
bridge sections with the same properties that are separated only by
whitespaces.

Related to issue #37

test: Test previous-output with zsh completion.

This change adds a test to cover the interaction of zsh completion and
prompt detection. This change relies on previous fix 8230077.

Related to issue #37