feat: Discover the start of multi-line prompt in zsh.
This change detects the special strings sent by zsh for the prompt_sp
option (enabled by default) and use that to mark the start of the
prompt. This is useful, as zsh enables bracketed paste only after
printing the prompt, so bracketed paste marks the end and not the
beginning of the prompt.

This change also turns the line continuation displayed when a command
output doesn't end with a newline by a real newline when a new prompt is
detected there. This looks a little bit better and guarantees that
prompts of shell that support prompt_sp always start at a real newline.

Issue #36