fix: Right prompt and zsh.
Before this change, right prompt detection wasn't working properly on
zsh.

The problem was that MisTTY assumed that we'd know we're in a
prompt (bracketed paste is on) when the right prompt was printed, but
that's not the case for zsh: the right prompt is printed with bracketed
paste still off and MisTTY realizes afterwards that that section was a
prompt.

This change fixes the issue by always setting mistty-maybe-skip to space
that were skipped, even if we're not in a prompt. It also sets
mistty-prompt-id to the line following the point where bracketed paste
was turned on.

Related to issue #36