test: Use a rc file for bash.
Before this change, on bash, the prompt was set interactively while on
fish and zsh, it was set at startup using a rc file. This change
switches batches to using a rc file as well, so shells work in similar
ways.

Additionally, using a rc file makes the startup procedure simpler and
allows testing what happens with the first prompt.