fix: Support first prompt refresh on zsh.
Zsh sometimes refreshes the prompt. Before this change, when that
happened on the first prompt, using a multi-line prompt, the start of
the prompt was lost and the result was handled as a single-line prompt.

This was due to MisTTY being confused about getting two
enable-bracketed-paste with no disable-bracketed-paste in between. This
change makes sure that extra enable or disable are just ignored.

Issue #36