fix: Fix issue with excessive kill-line with right prompt.
Before this change, a kill-line called to delete the line on a right
prompt would delete past that first line.

The issue was that kill-line being confused by
mistty--before-change-on-work, which would delete the right prompt after
kill-line decided the region to delete.

This change removes mistty--before-change-on-work to instead rely on
proper handling of mistty-skip. This change also moves the right prompt
in a replay, after the changes have been applied and before the refresh,
so to help sync make the right decisions, and put the pointer at the
right position.